refactor: store구조 정리/ util함수 정리
-store 상위에 도메인폴더를 제거
-비슷한 형식의 util함수는분리된 파일이 아닌  통합 관리
