--- src/service/api/pageable .ts
@@ -1,7 +1,0 @@
-export type SortType = 'like' | 'review';
-
-export interface PageableArg {
-  page: number;
-  size: number;
-  sort: SortType | null;
-}
