+++ src/service/api/pageable.ts
@@ -1,0 +1,7 @@
+export type SortType = 'like' | 'review';
+
+export interface PageableArg {
+  page: number;
+  size: number;
+  sort: SortType | null;
+}
