--- src/utils/format/formatDateToISOString.ts
@@ -1,8 +1,0 @@
-const formatDateToISOstring = (date: Date) => {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  return `${year}-${month}-${day}`;
-};
-
-export default formatDateToISOstring;
